+++ tools/test_gen_mame_neogeo_software.py
@@ -39,8 +39,31 @@ class MameSoftwareListTests(unittest.TestCase):
             )
 
             software_list = ET.parse(output).getroot()
-            software = software_list.find("./software[@name='smbneogeo']")
-            self.assertIsNotNone(software)
+            software_entries = software_list.findall("./software")
+            self.assertEqual(len(software_entries), 1)
+            software = software_entries[0]
+            self.assertEqual(software.get("name"), "smbneogeo")
+            self.assertNotIn("puzzledp", text)
+
+            actual_rom_parts = []
+            for dataarea in software.findall("./part/dataarea"):
+                for rom in dataarea.findall("rom"):
+                    actual_rom_parts.append(
+                        (
+                            dataarea.get("name"),
+                            rom.get("name"),
+                            int(rom.get("size", "0"), 0),
+                            int(rom.get("offset", "0"), 0),
+                            rom.get("loadflag"),
+                        )
+                    )
+            self.assertEqual(actual_rom_parts, list(mame_list.ROM_PARTS))
+
+            actual_area_sizes = {
+                dataarea.get("name"): int(dataarea.get("size", "0"), 0)
+                for dataarea in software.findall("./part/dataarea")
+            }
+            self.assertEqual(actual_area_sizes, mame_list.DATA_AREA_SIZES)
 
             maincpu = software.find("./part/dataarea[@name='maincpu']")
             self.assertIsNotNone(maincpu)